requirements.txt
llama-index
python-dotenv
html2text
